Distribute serialization tests over command modules
