+++ epp-client/src/extensions/rgp/report.rs
@@ -95,3 +95,61 @@ pub struct RgpRestoreReport {
     #[serde(rename = "rgp:restore", alias = "restore")]
     restore: RgpRestoreReportSection,
 }
+
+#[cfg(test)]
+mod tests {
+    use std::str::FromStr;
+
+    use chrono::DateTime;
+
+    use super::{RgpRestoreReport, Update};
+    use crate::domain::update::{DomainChangeInfo, DomainUpdate};
+    use crate::request::Transaction;
+    use crate::tests::{get_xml, CLTRID};
+
+    #[test]
+    fn command() {
+        let xml = get_xml("request/extensions/rgp_restore_report.xml").unwrap();
+
+        let pre_data =
+            "Pre-delete registration data goes here. Both XML and free text are allowed.";
+        let post_data =
+            "Post-restore registration data goes here. Both XML and free text are allowed.";
+        let deleted_at = DateTime::from_str("2021-07-10T22:00:00.0Z").unwrap();
+        let restored_at = DateTime::from_str("2021-07-20T22:00:00.0Z").unwrap();
+        let restore_reason = "Registrant error.";
+        let statements = &[
+        "This registrar has not restored the Registered Name in order to assume the rights to use or sell the Registered Name for itself or for any third party.",
+        "The information in this report is true to best of this registrar's knowledge, and this registrar acknowledges that intentionally supplying false information in this report shall constitute an incurable material breach of the Registry-Registrar Agreement.",
+    ];
+        let other = "Supporting information goes here.";
+
+        let domain_restore_report = Update {
+            data: RgpRestoreReport::new(
+                pre_data,
+                post_data,
+                deleted_at,
+                restored_at,
+                restore_reason,
+                statements,
+                other,
+            ),
+        };
+
+        let mut object = DomainUpdate::new("eppdev.com");
+        object.info(DomainChangeInfo {
+            registrant: None,
+            auth_info: None,
+        });
+
+        let serialized =
+            <DomainUpdate as Transaction<Update<RgpRestoreReport>>>::serialize_request(
+                object,
+                Some(domain_restore_report),
+                CLTRID,
+            )
+            .unwrap();
+
+        assert_eq!(xml, serialized);
+    }
+}
