+++ epp-client/src/host/delete.rs
@@ -49,7 +49,18 @@ mod tests {
     use crate::tests::{get_xml, CLTRID, SUCCESS_MSG, SVTRID};
 
     #[test]
-    fn host_delete() {
+    fn command() {
+        let xml = get_xml("request/host/delete.xml").unwrap();
+
+        let object = HostDelete::new("ns1.eppdev-1.com");
+
+        let serialized = object.serialize_request(None, CLTRID).unwrap();
+
+        assert_eq!(xml, serialized);
+    }
+
+    #[test]
+    fn response() {
         let xml = get_xml("response/host/delete.xml").unwrap();
         let object = HostDelete::deserialize_response(xml.as_str()).unwrap();
